moved fstop to modutil
