+++ src/api/libipc/src/modutil.c
@@ -67,3 +67,29 @@ int twaitpid(pid_t pid, int* status, int timeout) {
   // process did not exit
   return 0;
 }
+
+// Attempts to stop process using interrupt signal within timeout
+// If timeout exceeded and process has not terminated, SIGKILL is sent.
+static void fstop(pid_t pid, int sec_timeout, struct timespec retry_delay) {
+  bool pexited = false, psigterm = false;
+  time_t start, current, time_elapsed = 0;
+  time(&start);
+  while (time_elapsed < sec_timeout) {
+    kill(pid, SIGINT);
+    int stat;
+    waitpid(pid, &stat, WNOHANG);
+
+    pexited = WIFEXITED(stat);
+    psigterm = WIFSIGNALED(stat);
+    if (pexited || psigterm) break;
+
+    nanosleep(&retry_delay, NULL);
+    time(&current);
+    time_elapsed = current - start;
+  }
+
+  // Check to see if start process needs to be force stopped
+  if (!(pexited || psigterm)) {
+    kill(pid, SIGKILL);
+  }
+}
